+++ src/main/java/com/wintercogs/beyonddimensions/integration/module/curios/CuriosCapAttach.java
@@ -4,6 +4,7 @@ import com.wintercogs.beyonddimensions.common.init.BDItems;
 import com.wintercogs.beyonddimensions.common.item.NetFeederItem;
 import com.wintercogs.beyonddimensions.common.item.NetMagnetItem;
 import com.wintercogs.beyonddimensions.common.item.NetRestockerItem;
+import com.wintercogs.beyonddimensions.common.item.XpExchangeItem;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
@@ -109,6 +110,30 @@ public class CuriosCapAttach
                     }
                 },
                 BDItems.NET_RESTOCKER_ITEM
+        );
+        // 经验交换棒
+        evt.registerItem(
+                CuriosCapability.ITEM,
+                (stack, context) -> new ICurio()
+                {
+                    @Override
+                    public ItemStack getStack()
+                    {
+                        return stack;
+                    }
+
+                    @Override
+                    public void curioTick(SlotContext slotContext)
+                    {
+                        if (!(slotContext.entity().level() instanceof ServerLevel serverLevel)) return;
+                        if (stack.getItem() == BDItems.XP_EXCHANGE_ITEM.get())
+                        {
+                            XpExchangeItem item = (XpExchangeItem) stack.getItem();
+                            item.inventoryTick(stack, serverLevel, slotContext.entity(), null);
+                        }
+                    }
+                },
+                BDItems.XP_EXCHANGE_ITEM
         );
     }
 }