+++ src/main/java/com/wintercogs/beyonddimensions/integration/module/jade/BDJadePlugin.java
@@ -1,0 +1,24 @@
+package com.wintercogs.beyonddimensions.integration.module.jade;
+
+import com.wintercogs.beyonddimensions.common.block.NetedBlock;
+import com.wintercogs.beyonddimensions.common.block.entity.NetedBlockEntity;
+import snownee.jade.api.IWailaClientRegistration;
+import snownee.jade.api.IWailaCommonRegistration;
+import snownee.jade.api.IWailaPlugin;
+import snownee.jade.api.WailaPlugin;
+
+@WailaPlugin
+public class BDJadePlugin implements IWailaPlugin
+{
+    @Override
+    public void register(IWailaCommonRegistration registration)
+    {
+        registration.registerBlockDataProvider(NetedBlockNetworkProvider.INSTANCE, NetedBlockEntity.class);
+    }
+
+    @Override
+    public void registerClient(IWailaClientRegistration registration)
+    {
+        registration.registerBlockComponent(NetedBlockNetworkProvider.Client.INSTANCE, NetedBlock.class);
+    }
+}